Improved plot_halo
